Rough mode support
This puts in place the framework for selecting a
mode with various todos and warnings that need to
be addressed.

Remove Motivation from storing if it is core or non-core

Can now define a mode

Can now change into a mode when there are no items in the to do list

Gets past the importance system when making a do now list.

More calculation of urgency

Get the importance scope for an item when finding the parent

Partially implement get_category_by_urgency

Get things to the point of being able declare a new mode and state it

Remove some warnings

safekeeping