Improve usability
Change the order of items so the default item is to prioritize so if people are not reading closely they won't just do random items but by default prioritize items.

Also improve the text explaining when to state if something is a smaller item versus when something is blocked on a different thing.