Support copying DB
We now have support for copying the data in the database to a different user. This can be used to migrate or move the data from one machine or account to another or it can be used for testing to make a copy of the data and then do testing on it.

We also now give an error message if a parameter is wrong rather than silently ignore the param. This should help prevent mistakes or confusion if the program isn't working as expected.